requirements.txt
easyocr==1.4.1
Image==1.5.33
imageio==2.9.0
numpy==1.20.1
opencv_python_headless==4.5.3.56
Pillow==9.1.0
pipreqs==0.4.11
pytesseract==0.3.8
rfc3339==6.2
scikit_image==0.18.1
skimage==0.0
